+++ src/comp/ProfileCard/comp/ProfileHead.js
@@ -1,8 +1,8 @@
-import { Card, Avatar, List } from "antd";
+import { Card, Avatar } from "antd";
 
 const { Meta } = Card;
 
-const ProfileHead = () => {
+const ProfileHead = (props) => {
   return (
     <div>
       <Meta
@@ -12,7 +12,7 @@ const ProfileHead = () => {
             src="https://steamcdn-a.akamaihd.net/steamcommunity/public/images/avatars/ac/acd2539e1fab61c876e2883d028b94a0b1edece4_full.jpg"
           />
         }
-        title="Gamix"
+        title={props.name}
         description="Willkommen in der Member Area!"
       />
     </div>
